pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qvartools"
version = "0.0.0"
description = "Quantum variational toolkit for molecular ground-state energy estimation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "George Umbrarescu"},
]
keywords = [
    "quantum chemistry",
    "variational methods",
    "neural quantum states",
    "normalizing flows",
    "Krylov methods",
    "quantum diagonalization",
    "SQD",
    "SKQD",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "torch>=2.0",
]

[project.urls]
Homepage = "https://github.com/George930502/qvartools"
Documentation = "https://qvartools.readthedocs.io"
Repository = "https://github.com/George930502/qvartools"
Changelog = "https://github.com/George930502/qvartools/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/George930502/qvartools/issues"

[project.optional-dependencies]
pyscf = ["pyscf>=2.4"]
numba = ["numba>=0.57"]
gpu = ["cupy-cuda12x>=12.0"]
quantum = [
    "qiskit>=1.0",
    "qiskit-addon-sqd>=0.12",
    "ffsim>=0.0.70",
]
configs = ["pyyaml>=6.0"]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=2.0",
    "numpydoc>=1.6",
]
cudaq = ["cudaq>=0.14"]
full = ["pyscf>=2.4", "numba>=0.57"]
dev = ["pytest>=8", "pytest-cov>=5", "ruff>=0.4"]

[dependency-groups]
test = ["pytest>=8", "pytest-cov>=5"]
dev = [{include-group = "test"}, "ruff>=0.4", "mypy>=1.8"]

[tool.hatch.build.targets.wheel]
packages = ["src/qvartools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow",
    "gpu: requires GPU",
    "pyscf: requires PySCF",
]
filterwarnings = [
    # PyTorch internal: torch.jit.script_method deprecated in favour of torch.compile
    "ignore:.*torch\\.jit\\.script_method.*:DeprecationWarning",
    # cuquantum module reorganisation warnings (WSL2 environment)
    "ignore:.*custatevec module is deprecated.*:DeprecationWarning",
    "ignore:.*cutensornet module is deprecated.*:DeprecationWarning",
    "ignore:.*This module is deprecated and will be removed.*:DeprecationWarning",
]

[tool.ruff]
src = ["src"]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
